Remove LearningRoadmapSection component and its associated logic
